modified:   lib/codelog.rb 	renamed:    lib/codelog/presenters/version_data_presentator.rb -> lib/codelog/presenters/version_data_presenter.rb Changed the name of the file to match the name of the class.
