+++ lib/codelog.rb
@@ -8,4 +8,4 @@ require 'codelog/command/release'
 require 'codelog/command/step/changelog'
 require 'codelog/command/step/delete'
 require 'codelog/command/step/version'
-require 'codelog/presenters/version_data_presentator'
+require 'codelog/presenters/version_data_presenter'
