+++ lib/codelog/presenters/version_data_presenter.rb
